Modernize the Apollo Client testing approach
The current Apollo Client testing approach is based on a no
longer supported library. This PR helps address this by:

- Moving to the `@apollo/client` dependency (currently an RC version
  of 3.4)
- Switching the `react-apollo` naming for `apollo-client`
  (`react-apollo` has been deprecated)
- Simplifying the Apollo test application using Reactive Variables
  and our field policy API

Some of the Apollo Client tests still fail, but this is a stepping
stone in our React 18 work.
