+++ package.json
@@ -25,7 +25,7 @@
     "build:use-subscription": "cross-env NAME=use-subscription webpack",
     "build:react-state": "cross-env NAME=react-state webpack",
     "build:simplux": "cross-env NAME=simplux webpack",
-    "build:react-apollo": "cross-env NAME=react-apollo webpack",
+    "build:apollo-client": "cross-env NAME=apollo-client webpack",
     "build:recoil": "cross-env NAME=recoil webpack",
     "build:jotai": "cross-env NAME=jotai webpack",
     "build:effector": "cross-env NAME=effector webpack",
@@ -40,11 +40,10 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "@apollo/react-hooks": "^4.0.0",
+    "@apollo/client": "^3.4.0-rc.14",
     "@react-rxjs/core": "^0.8.2",
     "@simplux/core": "^0.17.0",
     "@simplux/react": "^0.17.0",
-    "apollo-boost": "^0.4.9",
     "constate": "^3.2.0",
     "effector": "^21.8.11",
     "effector-react": "^21.3.2",
