+++ webpack.config.js
@@ -10,6 +10,9 @@ module.exports = {
     filename: '[name].js',
   },
   plugins: [
+    new webpack.DefinePlugin({
+      __DEV__: JSON.stringify(process.env.NODE_ENV !== "production"),
+    }),
     new webpack.EnvironmentPlugin(['NAME']),
     new HtmlWebpackPlugin({
       template: './public/index.html',
